dags/pipelines/nasa_apod/requirements.txt
requests
pandas
python-dotenv  
s3fs
sqlalchemy
pyscopg2-binary
minio
